Move the Luminosity grid computation on the run analysis action.
git-svn-id: http://euclid.esac.esa.int/svn/EC/SGS/SDC/CH/Projects/Phosphoros/branches/Luminosity@10257 463500fe-d6be-4184-9b20-47717e75ecac

+++ PhzQtUI/src/lib/DialogLuminosityPrior.cpp
@@ -26,21 +26,11 @@
 #include "PhzQtUI/DialogZRanges.h"
 #include "PhzQtUI/DialogFilterSelector.h"
 
-#include "PhzModeling/SparseGridCreator.h"
-#include "PhzModeling/NoIgmFunctor.h"
 
 #include "Configuration/ConfigManager.h"
 #include "DefaultOptionsCompleter.h"
-#include "PhzConfiguration/ComputeModelGridConfig.h"
-#include "PhzConfiguration/SedProviderConfig.h"
-#include "PhzConfiguration/ReddeningProviderConfig.h"
-#include "PhzConfiguration/FilterProviderConfig.h"
-#include "PhzConfiguration/LuminosityBandConfig.h"
-#include "PhzConfiguration/IgmConfig.h"
-#include "PhzConfiguration/ModelGridOutputConfig.h"
-#include "PhzConfiguration/ComputeLuminosityModelGridConfig.h"
 
-using namespace Euclid::PhzConfiguration;
+
 
 namespace Euclid {
 namespace PhzQtUI {
@@ -205,10 +195,6 @@ void DialogLuminosityPrior::on_btn_save_clicked(){
     return;
   }
 
-  // switch to the progress bar
-  manageBtnEnability(false, true);
-  ui->stackedWidget->setCurrentIndex(1);
-
   auto old_grid_name = info.getLuminosityModelGridName();
 
   updateInfo(info);
@@ -232,17 +218,16 @@ void DialogLuminosityPrior::on_btn_save_clicked(){
   stream<<xml;
   file.close();
 
+  auto new_grid_name = info.getLuminosityModelGridName();
+
   // model grid file handling
-  if (old_grid_name.length()>0){
+  if (old_grid_name.length()>0 && new_grid_name!= old_grid_name){
     QFile::remove(m_grid_folder+QDir::separator()+QString::fromStdString(old_grid_name));
   }
 
-  computeModelGrid(info);
-  ui->stackedWidget->setCurrentIndex(0);
   manageBtnEnability(false, false);
 
   m_new=false;
-
 }
 
 void DialogLuminosityPrior::on_btn_cancel_clicked(){
@@ -681,66 +666,7 @@ void DialogLuminosityPrior::updatePriorRow(QModelIndex& index,const size_t& row,
 }
 
 
-class ProgressReporter {
-public:
 
-  ProgressReporter(DialogLuminosityPrior* dialog){
-    m_dialog=dialog;
-  }
-
-  void operator()(int current, int total) {
-    m_dialog->onProgress(current,total);
-  }
-private:
-  DialogLuminosityPrior* m_dialog;
-};
-
-
-
-void DialogLuminosityPrior::computeModelGrid(const LuminosityPriorConfig& info ){
-  std::map<std::string, boost::program_options::variable_value> options_map;
-  completeWithDefaults<ComputeLuminosityModelGridConfig>(options_map);
-
-  for (auto&pair : FileUtils::getPathConfiguration(false,true,true,false)){
-    options_map[pair.first]=pair.second;
-  }
-
-  for (auto& pair : info.getBasicConfigOptions(false)){
-    options_map[pair.first]=pair.second;
-  }
-
-  options_map["catalog-type"].value() = boost::any(std::string(m_survey_name));
-
-
-
-  options_map["model-grid-file"].value() = boost::any(m_model_grid_name);
-
-  options_map["output-model-grid"].value() = boost::any(info.getLuminosityModelGridName());
-
-  long config_manager_id = std::chrono::duration_cast<std::chrono::microseconds>(
-                                     std::chrono::system_clock::now().time_since_epoch()).count();
-  auto& config_manager = Configuration::ConfigManager::getInstance(config_manager_id);
-  config_manager.registerConfiguration<ComputeLuminosityModelGridConfig>();
-  config_manager.closeRegistration();
-  config_manager.initialize(options_map);
-
-  auto& sed_provider = config_manager.getConfiguration<SedProviderConfig>().getSedDatasetProvider();
-  auto& reddening_provider = config_manager.getConfiguration<ReddeningProviderConfig>().getReddeningDatasetProvider();
-  const auto& filter_provider = config_manager.getConfiguration<FilterProviderConfig>().getFilterDatasetProvider();
-  auto& igm_abs_func = config_manager.getConfiguration<IgmConfig>().getIgmAbsorptionFunction();
-
-  Euclid::PhzModeling::SparseGridCreator creator {
-               sed_provider, reddening_provider, filter_provider, igm_abs_func};
-
-  auto param_space_map = config_manager.getConfiguration<ComputeLuminosityModelGridConfig>().getParameterSpaceRegions();
-  std::vector<Euclid::XYDataset::QualifiedName> filter_list = {config_manager.getConfiguration<Euclid::PhzConfiguration::LuminosityBandConfig>().getLuminosityFilter()};
-
-  auto results = creator.createGrid(param_space_map, filter_list, ProgressReporter{this});
-
-  dialog_logger.info() << "Creating the grid output";
-  auto output = config_manager.getConfiguration<ModelGridOutputConfig>().getOutputFunction();
-  output(results);
-}
 
 }
 }
